requirements.txt
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.9
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
